requirements.txt
# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Windows process/window tracking
psutil>=5.9.0
pywin32>=306

# Scheduler
apscheduler>=3.10.0

# Slack
slack-sdk>=3.26.0
requests>=2.31.0

# OpenAI
openai>=1.12.0

# Tray icon
pystray>=0.19.0
Pillow>=10.0.0

# Web UI
flask>=3.0.0
